+++ lib/src/hashlib_base.dart
@@ -1,3 +1,4 @@
 export 'package:hashlib/src/md5.dart';
 export 'package:hashlib/src/sha1.dart';
+export 'package:hashlib/src/sha224.dart';
 export 'package:hashlib/src/sha256.dart';