+++ lib/src/md5.dart
@@ -56,22 +56,10 @@ class MD5 extends HashAlgo {
   /// Initializes a new instance of MD5 message-digest.
   /// An instance can be re-used after calling the [clear] function.
   MD5() {
-    clear();
-  }
-
-  @override
-  void clear() {
-    // Reset count
-    _countLow = 0;
-    _countHigh = 0;
-    // Initialize variables
     _state[0] = 0x67452301; // a
     _state[1] = 0xefcdab89; // b
     _state[2] = 0x98badcfe; // c
     _state[3] = 0x10325476; // d
-    // Reset state
-    _pos = 0;
-    _closed = false;
   }
 
   @override